more rename of lemma_lattices to lattices
