--- lemma_lattices/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class LemmaLatticesConfig(AppConfig):
-    name = "lemma_lattices"
